First commit, works, need more styling
Still working on these! if you download and save them to a folder on your xampp, start with index.php.

I'm pulling the questions from the questions.txt file. The questions were just for testing with, we can use the ones Nate has made. I am still working on saving the actual user score to the scores file.